requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.2.3
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-filter==25.1
django-import-export==4.3.9
django-phonenumber-field==8.1.0
django-simple-history==3.10.1
django-tinymce==4.1.0
django-trench==0.3.1
django-unfold==0.63.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-extra-fields==3.7.0
drf-yasg==1.21.10
filetype==1.2.0
frozenlist==1.7.0
idna==3.10
inflection==0.5.1
Markdown==3.8.2
multidict==6.6.3
oauthlib==3.3.1
packaging==25.0
phonenumbers==9.0.11
pillow==11.3.0
propcache==0.3.2
psycopg==3.2.9
pycparser==2.22
PyJWT==2.9.0
pyotp==2.9.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
smsapi-client==2.9.6
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tablib==3.8.0
twilio==9.6.4
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.20.1
yubico-client==1.13.0
